HYD-4859 Realtime@3.0.4: Add alerts to long-polling list
The latest long polling patch added alerts to endpoints that can be
polled. As such, we need to update realtime to account for all /alert
requests to delegate to long polling.

While were at it, we should extract the long polling logic into it's own
module, so we can utilize it in the custom /health endpoint, which uses
alerts to derive it's state.

  - Extract long-polling logic.
  - Update /health route to use fp.
  - Remove lodash-mixins.
  - Add clone as shim for obj implementation.
  - Update /alert to be long-polled
    in wildcard and health routes.
  - Add tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>