HYD-5096 Update Stub Daddy
The realtime test is seeing a race condition in which we expect a call
to be made once but stub daddy is claiming that it is being called
twice. This is an intermittent error and does not always occur, but it
is much more prone to occur on CI. The purpose of this ticket is to
identity the cause of the race condition in stub daddy and fix it.

With the new updates in stub daddy, this patch will fix the re-poll test race
condition occurring in the wildcard test. I've also updated the other
integration tests to work with the new stub daddy format.

Signed-off-by: Will Johnson <william.c.johnson@intel.com>