HYD-5464 Realtime: add info level logging that makes api calls viewable
We currently have a display breakdown somewhere in the
API -> Realtime -> GUI pipeline where data is not being updated properly. There are some possible causes:

1) API is not sending changes.
2) Stream of data is drying up somewhere.
3) GUI is not digesting changes to display.

We currently lack logging to display the clear flow of information through the system.
Once we have this, we can narrow in on a specific area of the app
causing these issues and focus our fixes there.

This ticket will associate all sockets / requests with an id so we can trace their lifetimes correctly though the system.

  - Create serializer for sockets and socket requests.
  - Add info logging to request setup / routing / teardown.
  - Remove unit testing of logging.
  - Switch from jshint to eslint.
  - Bump some deps.
  - Update copyright to 2016.

Signed-off-by: Joe Grund <joe.grund@intel.com>