HYD-6743 Add session route to realtime
The remote validate changes require that the login screen be updated to
use streams instead of promises. This allows us to remove the old
session model that is no longer needed. The problem we run into is that
when we make an http request directly, a "set-cookie" header is included
in the response, which sets the new session id. Once the session id
cookie is set and we refresh the page, we use the new session and the
user is logged in. Since we are using a socket now instead of the
session model, we simply get a response of null back from the session
post and the "set-cookie" header is never returned to the app. This is a
problem because the session id never gets set and thus the user never
gets logged in.

To resolve this, a new "session" route can be added to realtime that
will:

1. Return the "set-cookie" header instead of null
2. Update the session id in realtime

The web app can then set the cookie appropriately and realtime will now
use the updated session id for every future call.

This patch will:

1. Create a session route
2. Add the session route

Signed-off-by: Will Johnson <william.c.johnson@intel.com>