Up to date for now (#4)
* HYD-6743 Add session route to realtime

The remote validate changes require that the login screen be updated to
use streams instead of promises. This allows us to remove the old
session model that is no longer needed. The problem we run into is that
when we make an http request directly, a "set-cookie" header is included
in the response, which sets the new session id. Once the session id
cookie is set and we refresh the page, we use the new session and the
user is logged in. Since we are using a socket now instead of the
session model, we simply get a response of null back from the session
post and the "set-cookie" header is never returned to the app. This is a
problem because the session id never gets set and thus the user never
gets logged in.

To resolve this, a new "session" route can be added to realtime that
will:

1. Return the "set-cookie" header instead of null
2. Update the session id in realtime

The web app can then set the cookie appropriately and realtime will now
use the updated session id for every future call.

This patch will:

1. Create a session route
2. Add the session route

Signed-off-by: Will Johnson <william.c.johnson@intel.com>

* - Added more tests and small route updates

* - Cleaned up processSession
- Removed GET session route. With the updated router this will do fine
  going to wildcard-routes

* - more cleanup

* - removed getRequest and call apiRequest directly in the handlers.

* More updates and adding test for session DELETE

* - removed jscs
- fixed eslint error

* - Abstracted sessionRoute function into a single location

* More cleanups

* remove node 0.10 and add 6

* - Update version to v5.1.0

* HYD-7080 Realtime: bump deps

Bump realtime deps to latest.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* HYD-7129 Remove node-clone

Remove node-clone which is on the SDL blacklist.

We should be able to replace with es6 destructuring + obj.clone

  - Remove node-clone.

Signed-off-by: Joe Grund <joe.grund@intel.com>

* Update run_test.sh

* Adding .travis.yml file

* Remove

* Update to the minimal set of deps for a
pass.

This module will be retired in some form.

* Update all copyrights again.

* Add .npmignore.