Properly wait for dependencies
This patch adds an `ENTRYPOINT` to wait for postgres + IML conf to be
available before starting realtime.

This avoids the situation where realtime get's hung up waiting for
Postgres to be available at the same time Postgres is starting.

This could be thought as a poor-mans analouge to systemd `After=`.

This patch also adds ENV vars for the db and realtime port. These can be
overriden, but there is not really a point to doing so in a
containerized app.

It also adds `dumb-init` so node does not run as pid 1 and can better
handle signals like `SIGTERM`.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>