Consolidate error handling
This patch consolidates error handling by allowing all errors to
propagate up to `unhandledRejection` + `uncaughtException` handlers.

When either of these handlers are hit we stop socket.io, end the pool,
and close all connection sockets.

By doing this, we clear the event loop and ensure node will exit. We
also allow for a central error handling point and to fail fast on
issues.

Signed-off-by: Joe Grund <jgrund@whamcloud.io>